Create process_snap_data.py
This code processes hourly SNAP data holdings for the variables tsk, u10, v10, seaice, and psfc. The data is read in and processed one variable at a time, with backup files being saved after each variable has successfully been processed. The code reads in the data, crops it to the data nearest to our target region, downscales it to daily measurements, then interpolates it day-by-day on input transect lines. It outputs a data frame containing daily measurements of each input variable for each year on each transect. It also contains calculated wind speed and wind direction for each transect and day that both u10 and v10 components are present.