Enhance some function , make it more general
I replace RE_format, PS_format with get_image_date. This function is more general to work with any satellite.

When I request Rapid eye and planet scope satellite, I found that they have the same id format. But the functions written before suppose that they are different.

I make this function take the date format and date length to avoid all these to avoid  differences. If there are differences in the future we will  update only the variables with no need to change all function.